ron

Module de

Re-exports

pub use crate::error::Error;

pub use crate::error::Position;

pub use crate::error::SpannedError;

Modules

id 🔒

tag 🔒

value 🔒

Macros

guard_recursion 🔒

Structs

CommaSeparated 🔒

Deserializer

The RON deserializer.

Enum 🔒

SerdeEnumContent 🔒

Enums

Terminator 🔒

Constants

SERDE_CONTENT_CANARY 🔒

SERDE_TAG_KEY_CANARY 🔒

Functions

from_bytes

A convenience function for building a deserializer and deserializing a value of type T from bytes.

from_reader

A convenience function for building a deserializer and deserializing a value of type T from a reader.

from_str

A convenience function for building a deserializer and deserializing a value of type T from a string.

struct_error_name 🔒