Struct ron::de::Deserializer

pub struct Deserializer<'de> {
    bytes: Bytes<'de>,
    newtype_variant: bool,
    last_identifier: Option<&'de str>,
    recursion_limit: Option<usize>,
}

The RON deserializer.

If you just want to simply deserialize a value, you can use the from_str convenience function.

Fields
bytes: Bytes<'de>newtype_variant: boollast_identifier: Option<&'de str>recursion_limit: Option<usize>
Implementations

impl<'de> Deserializer<'de>

pub fn from_str(input: &'de str) -> SpannedResult<Self>

pub fn from_bytes(input: &'de [u8]) -> SpannedResult<Self>

pub fn from_str_with_options( input: &'de str, options: Options ) -> SpannedResult<Self>

pub fn from_bytes_with_options( input: &'de [u8], options: Options ) -> SpannedResult<Self>

pub fn remainder(&self) -> Cow<'_, str>

pub fn span_error(&self, code: Error) -> SpannedError

impl<'de> Deserializer<'de>

pub fn end(&mut self) -> Result<()>

Check if the remaining bytes are whitespace only, otherwise return an error.

fn handle_any_struct<V>(&mut self, visitor: V) -> Result<V::Value>where V: Visitor<'de>,

Called from deserialize_any when a struct was detected. Decides if there is a unit, tuple or usual struct and deserializes it accordingly.

This method assumes there is no identifier left.

fn handle_struct_after_name<V>( &mut self, name_for_pretty_errors_only: &'static str, visitor: V ) -> Result<V::Value>where V: Visitor<'de>,

Called from deserialize_struct, struct_variant, and handle_any_struct. Handles deserialising the enclosing parentheses and everything in between.

This method assumes there is no struct name identifier left.