Struct ron::ser::Serializer

pub struct Serializer<W: Write> { /* private fields */ }

The RON serializer.

You can just use to_string for deserializing a value. If you want it pretty-printed, take a look at the pretty module.

Implementations

impl<W: Write> Serializer<W>

pub fn new(
writer: W,
config: Option<PrettyConfig>,
struct_names: bool
) -> Result<Self>

๐Deprecated: Serializer::new is deprecated because struct_names was moved to PrettyConfig

Creates a new Serializer.

Most of the time you can just use to_string or to_string_pretty.

Deprecation

This constructor takes struct_names, which has been moved to PrettyConfig. To stay semver compatible the Serializer will keep the struct_names field until the next minor version gets released, and struct names will be generated if either the PrettyConfigโs or the Serializerโs struct name field is true.

pub fn with_options(
writer: W,
config: Option<PrettyConfig>,
options: Options
) -> Result<Self>

Creates a new Serializer.

Most of the time you can just use to_string or to_string_pretty.