ron::value

Struct RawValue

pub struct RawValue { /* private fields */ }

Implementations

impl RawValue

pub fn get_ron(&self) -> &str

Get the inner raw RON string, which is guaranteed to contain valid RON.

pub fn from_ron(ron: &str) -> SpannedResult<&Self>

Helper function to validate a RON string and turn it into a RawValue.

pub fn from_boxed_ron(ron: Box<str>) -> SpannedResult<Box<Self>>

Helper function to validate a RON string and turn it into a RawValue.

pub fn into_rust<'de, T: Deserialize<'de>>(&'de self) -> SpannedResult<T>

Helper function to deserialize the inner RON string into T.

pub fn from_rust<T: Serialize>(value: &T) -> Result<Box<Self>, Error>

Helper function to serialize value into a RON string.

impl RawValue

pub fn trim(&self) -> &Self

Trims any leadning and trailing whitespace off the raw RON string, including whitespace characters and comments.

pub fn trim_boxed(self: Box<Self>) -> Box<Self>

Trims any leadning and trailing whitespace off the boxed raw RON string, including whitespace characters and comments.