Module ron::de

Structs

Deserializer
The RON deserializer.

Position

Enums

Error

ParseError

Functions

from_bytes
A convenience function for building a deserializer and deserializing a value of type T from bytes.

from_reader
A convenience function for reading data from a reader and feeding into a deserializer.

from_str
A convenience function for building a deserializer and deserializing a value of type T from a string.

Type Definitions

Result
Deserialization result.