Module ron::de

Re-exports

pub use crate::error::Error;

pub use crate::error::ErrorCode;

pub use crate::error::Result;

Structs

Deserializer
The RON deserializer.

Position

Functions

from_bytes
A convenience function for building a deserializer and deserializing a value of type T from bytes.

from_reader
A convenience function for reading data from a reader and feeding into a deserializer.

from_str
A convenience function for building a deserializer and deserializing a value of type T from a string.