Struct ron::de::Deserializer

pub struct Deserializer<'de> { /* fields omitted */ }

The RON deserializer.

If you just want to simply deserialize a value, you can use the from_str convenience function.

Implementations

impl<'de> Deserializer<'de>

pub fn from_str(input: &'de str) -> Result<Self>

pub fn from_bytes(input: &'de [u8]) -> Result<Self>

pub fn remainder(&self) -> Cow<str>

impl<'de> Deserializer<'de>

pub fn end(&mut self) -> Result<()>

Check if the remaining bytes are whitespace only, otherwise return an error.