ron::de

Struct Deserializer

pub struct Deserializer<'de> { /* private fields */ }

The RON deserializer.

If you just want to simply deserialize a value, you can use the from_str convenience function.

Implementations

impl<'de> Deserializer<'de>

pub fn from_str(input: &'de str) -> SpannedResult<Self>

pub fn from_bytes(input: &'de [u8]) -> SpannedResult<Self>

pub fn from_str_with_options( input: &'de str, options: &Options, ) -> SpannedResult<Self>

pub fn from_bytes_with_options( input: &'de [u8], options: &Options, ) -> SpannedResult<Self>

pub fn remainder(&self) -> &'de str

pub fn span_error(&self, code: Error) -> SpannedError

pub fn extensions(&self) -> Extensions

impl<'de> Deserializer<'de>

pub fn end(&mut self) -> Result<()>

Check if the remaining bytes are whitespace only, otherwise return an error.